// client/dist/widgets/arcgis/near-me/src/config.ts
export interface Config {
  generalSettings: GeneralSettings
  configInfo: any
}

export interface ConfigInfo {
  [dataSourceId: string]: IndividualConfigInfo
}

interface IndividualConfigInfo {
  analysisSettings: AnalysisSettings
  searchSettings: SearchSettings
}

export interface GeneralSettings {
  highlightColor: string
  noResultsFoundText: string
  noResultMsgStyleSettings: FontStyleSettings
  promptTextMessage: string
  promptTextMsgStyleSettings: FontStyleSettings
}
// ...
export const enum AnalysisTypeName {
  Closest = 'closest',
  Proximity = 'proximity',
  Summary = 'summary'
}

export interface SummaryAttributes {
  [key: string]: any
}

export type IMConfig = ImmutableObject<Config>